% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_png_plot.R
\name{save_png_plot}
\alias{save_png_plot}
\title{Save plot in .png}
\usage{
save_png_plot(
  plot,
  name = "test",
  folder = "./",
  type = "standard",
  format = "svg",
  width = 480,
  height = 480,
  pointsize = 12
)
}
\arguments{
\item{plot}{A variable of class \code{recordedplot} or \code{ggplot}, see examples.}

\item{name}{The name of the plot without extensions. Default is \code{"test"}.}

\item{folder}{The path of the folder where to save the plot. Default is the current folder.}

\item{type}{\code{"standard"} if plot is a standard R plot or \code{"ggplot"} if plot is a ggplot.}

\item{width}{the width of the device.}

\item{height}{the height of the device.}

\item{pointsize}{the default pointsize of plotted text.}
}
\description{
A function that allows you to save standard R plots
or ggplot plots in .png format.
}
\examples{
data = data.frame(x = c(1, 2, 3, 4, 5), y = c(2, 4, 6, 8, 10))

### ggplot example
library(ggplot2)
plot1 = ggplot(data = data, aes(x = x, y = y)) + geom_point()
save_png_plot(
  plot1,
  name = "ggplot_test",
  type = "ggplot"
)

### standard plot example
plot(data$x,data$y, xlab="x", ylab="y",pch=19)
points(3,5,col="red",pch=19)
plot2 = recordPlot()
save_png_plot(
  plot2,
  name = "standard_plot_test",
  type = "standard"
)
}
